refactor: merge edit/read/create Note modals
