+++ backend/src/main/resources/i18n/fr.json
@@ -38,7 +38,9 @@
   "collaborativewall.modal.modify": "Modifier",
   "collaborativewall.modal.save": "Enregistrer",
   "collaborativewall.modal.add": "Ajouter",
-  "collaborativewall.modal.create": "Nouvelle Note",
+  "collaborativewall.modal.title.read": "Note",
+  "collaborativewall.modal.title.edit": "Modification",
+  "collaborativewall.modal.title.create": "Nouvelle Note",
   "collaborativewall.note.notfound": "Note introuvable.",
   "collaborativewall.color.white": "Blanc",
   "collaborativewall.color.yellow": "Jaune",
