+++ frontend/src/features/note-modal/components/ColorSelect.tsx
